fixed account login and signup expanding to overflow the header area
Now the login form blends in seamlessly with the header area and the header has been sized to a minimum of 40 px and the account div to a max of 40 px.
This allows the header to expand as needed but the account div to not overflow the minimum height; which takes in regard current one line header style metric.